Revamp and update README (#137)
A lot of information in the README was out of date, and I've also seen lots of people have trouble sorting through it at meetups. With these changes, I'm hoping to offer a clearer overview and more direct access to all the sub-projects where most of the work happens.

I've dropped the "technologies used" list because it's so broad it doesn't feel especially helpful. I've tried to preserve some flavor of it by including some of the major technologies/languages in the list of sub-projects.