Add maintenance status notice (#169)
This is part of #168. EDGI is no longer making active use of the tools here, and they never became generalized to the point where they aren’t a huge amount of effort to deploy and maintain by other organizations or individuals (for example, you need a close partnership with the Internet Archive or another organization that crawls/scrapes/archives the monitored URLs).

The goal here is to make the status of things clear and provide some useful resources for anybody looking at this project.