Fix deprecation warning MD syntax
The special markdown syntax GitHub is using for warning callouts has changed; this updates it to the new syntax.

This *should* work, but apparently the current version does not work in previews (when editing or in PR diffs), so you’ll need to view the file in this branch to see if it’s working correctly.

It’s also possible I should just switch this to a non-specialized syntax. Since the specialized syntax here is non-standard and still marked “beta”, it might not be good to rely on (but it has been supported for over a year, so… 🤷 ).